# Description

If 'Last-Modified' and 'If-Modified-Since' are equal, then `DefaultCtx.Fresh` returns false.
However, if no updates have been made, the cache should not to be stale.
Thank you.

## Type of change

- [x] Enhancement (improvement to existing features and functionality)

## Checklist

- [ ] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [x] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [x] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [x] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [ ] Aimed for optimal performance with minimal allocations in the new code.
- [x] Provided benchmarks for the new code to analyze and improve upon.

## Benchmarks(on GitHub Codespaces)

```sh
go test -v -run=^$ -bench=Benchmark_Ctx_Fresh_LastModified -benchmem -count=4
```


### Before
```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD EPYC 7763 64-Core Processor                
Benchmark_Ctx_Fresh_LastModified
Benchmark_Ctx_Fresh_LastModified-2      10880254               123.4 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_Fresh_LastModified-2      10743270               109.9 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_Fresh_LastModified-2      10766181               114.0 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_Fresh_LastModified-2      10719219               108.6 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v3     6.345s
```

### After
```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD EPYC 7763 64-Core Processor                
Benchmark_Ctx_Fresh_LastModified
Benchmark_Ctx_Fresh_LastModified-2      10672034               110.4 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_Fresh_LastModified-2      10670136               126.8 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_Fresh_LastModified-2      10492482               114.3 ns/op             0 B/op          0 allocs/op
Benchmark_Ctx_Fresh_LastModified-2      10615524               114.2 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v3     5.414s
```